+++ openeo_driver/constants.py
@@ -44,6 +44,8 @@ RESAMPLE_SPATIAL_METHODS = [
     "q3",
     "rms",
     "sum",
+    "binning", # custom method for "binning" resampling of sentinel-3
+    "geocode" #custom method for geocoding of e.g. Sentinel-1 based on lat/lon bands
 ]
 
 # Align options as used in official spec of `resample_spatial`